Changed the way transactions work.
Only one transaction is ever started.  Nested transaction simply
increment a counter.  When commiting, if an error was detected, a flag
is set and when the counter drops back to 1, a real commit or rollback
is performed.  If a rollback fails, it is a fatal error.  When a
rollback is performed, cleanup functions can be called.  This allows
inner "transactions" to register a cleanup function that is called if
the real transaction fails.